% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataSet}
\alias{DataSet}
\alias{LUAD_RESLIST3}
\alias{LUSC_RESLIST3}
\alias{ESCA_RESLIST3}
\title{Datasets
LUAD_RESLIST3 contains gene annot info}
\description{
Datasets
LUAD_RESLIST3 contains gene annot info

Datasets
LUSC_RESLIST3 contains gene annot info

Datasets
ESCA_RESLIST3 contains gene annot info
}
\keyword{datasets}
